+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Show whether VSync is enabled
 - Fix HUD alignment of negative three-digit angles
 - [Windows] Fix generation of timestamps in logs
+- Can now explore the Kerbol system
 
 # v0.9 - 2020-01-09
 - Output log to file (last.log)